refactor : 아티클 리펙토링 구조
