+++ src/main/java/com/sandbox/domain/articles/dto/ArticleResp.java
@@ -1,0 +1,4 @@
+package com.sandbox.domain.articles.dto;
+
+public class ArticleResp {
+}
