+++ src/main/java/com/sandbox/domain/articles/entity/Article.java
@@ -1,15 +1,15 @@
-package com.sandbox.domain.articles.dto;
+package com.sandbox.domain.articles.entity;
 
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDateTime;
-
 @Entity
 @Getter @Setter
 @NoArgsConstructor
+@AllArgsConstructor
 public class Article {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
