+++ src/main/java/com/sandbox/domain/articles/exception/ArticleExceptionHandler.java
@@ -1,0 +1,4 @@
+package com.sandbox.domain.articles.exception;
+
+public class ArticleExceptionHandler {
+}
