+++ src/main/java/com/sandbox/domain/articles/exception/ErrorArticleResp.java
@@ -1,0 +1,4 @@
+package com.sandbox.domain.articles.exception;
+
+public class ErrorArticleResp extends RuntimeException{
+}
